功能: 添加历史记录过期时间；下载远程时，如果本地历史记录存在则无需下载

+++ src/SyncClipboard.Core/Utilities/Job/AppdataFileDeleteJob.cs
@@ -28,7 +28,7 @@ public class AppdataFileDeleteJob(ConfigManager configManager) : IJob
                 foreach (var dirs in tempFolders)
                 {
                     var isTime = DateTime.TryParseExact(dirs.Name, "yyyyMMdd", CultureInfo.InvariantCulture, DateTimeStyles.None, out var createTime);
-                    if (!isTime || ((DateTime.Today - createTime) > TimeSpan.FromDays(config.TempFileRemainDays)))
+                    if (isTime && ((DateTime.Today - createTime) > TimeSpan.FromDays(config.TempFileRemainDays)))
                     {
                         dirs.Delete(true);
                     }
